189 - template query impl + test
