+++ dookug-common/dookug-common-util/src/main/java/hu/icellmobilsoft/dookug/common/util/date/DateUtil.java
@@ -19,11 +19,47 @@
  */
 package hu.icellmobilsoft.dookug.common.util.date;
 
+import java.time.LocalTime;
+import java.time.OffsetDateTime;
+
 /**
  * Date util.
  *
  * @author imre.scheffer
  */
 public class DateUtil extends hu.icellmobilsoft.coffee.tool.utils.date.DateUtil {
+
+    /**
+     * Start of day: 00:00 at the same offset
+     *
+     * @param odt
+     *            offset date time
+     * @return start of day
+     */
+    public static OffsetDateTime startOfDay(OffsetDateTime odt) {
+        return odt.toLocalDate().atStartOfDay().atOffset(odt.getOffset());
+    }
+
+    /**
+     * End of day (inclusive upper bound): same day at 23:59:59.999999999 at the same offset
+     *
+     * @param odt
+     *            offset date time
+     * @return end of day
+     */
+    public static OffsetDateTime endOfDayInclusive(OffsetDateTime odt) {
+        return odt.toLocalDate().atTime(LocalTime.MAX).atOffset(odt.getOffset());
+    }
+
+    /**
+     * End of day (exclusive upper bound): start of next day at the same offset
+     *
+     * @param odt
+     *            offset date time
+     * @return end of day exclusive
+     */
+    public static OffsetDateTime endOfDayExclusive(OffsetDateTime odt) {
+        return startOfDay(odt).plusDays(1);
+    }
 
 }
